refactor: rename delivery packages to handler and update main entry point to cmd/rest
