+++ internal/application/health/handler/rest/health.handler.go
@@ -1,4 +1,4 @@
-package http
+package rest
 
 import (
 	"net/http"
