+++ internal/application/health/handler/rest/health.handler_test.go
@@ -1,4 +1,4 @@
-package http
+package rest
 
 import (
 	"errors"
