+++ internal/application/order/handler/rest/order.handler.go
@@ -1,4 +1,4 @@
-package http
+package rest
 
 import (
 	"github.com/gin-gonic/gin"
